Renamed NewsletterDetailModalLabs to Prototype
no issue

- naming better reflects it's usage as a non-wired prototype and reduces confusion alongside the `NewsletterDetailsModalAlpha` file
