+++ apps/admin-x-settings/src/components/settings/email/newsletters/NewsletterDetailModalPrototype.tsx
